Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.entities.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service  //para registrar como componente/service Spring
@@ -29,9 +31,25 @@ public class UserService {
 		return repository.save(obj); //por padrão ele já retorna o objeto salvo
 	}
 	
+	/*public void delete(Long id) {
+		try {
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		}
+	}*/
+	
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id); // entity objeto monitorado pelo JPA
